Make the feedback form more configurable and feedbacks are saved in wiki pages
This commit brings a few improvements to the feedback. Users can provide there
own feedback forms, validation code and feedbacks are saved in wiki pages.
Some features are temporarily missing: feedbacks are not sent by mail for now
and saving in the wiki cannot be disabled. This is coming in the next commits.
Documentation is also coming.